Fix bugs related to no using live-base image as a base anymore

+++ runner/dl_checkpoints.sh
@@ -91,8 +91,11 @@ function download_live_models() {
     huggingface-cli download KBlueLeaf/kohaku-v2.1 --include "*.safetensors" "*.json" "*.txt" --exclude ".onnx" ".onnx_data" --cache-dir models
     huggingface-cli download stabilityai/sd-turbo --include "*.safetensors" "*.json" "*.txt" --exclude ".onnx" ".onnx_data" --cache-dir models
 
-    # ComfyUI
-    # docker pull $AI_RUNNER_COMFYUI_IMAGE
+    # ComfyUI models
+    if ! docker image inspect $AI_RUNNER_COMFYUI_IMAGE >/dev/null 2>&1; then
+        echo "ERROR: ComfyUI base image $AI_RUNNER_COMFYUI_IMAGE not found"
+        return 1
+    fi
     # ai-worker has tags hardcoded in `var livePipelineToImage` so we need to use the same tag in here:
     docker image tag $AI_RUNNER_COMFYUI_IMAGE livepeer/ai-runner:live-app-comfyui
     docker run --rm -v ./models:/models --gpus all -l ComfyUI-Setup-Models $AI_RUNNER_COMFYUI_IMAGE \
@@ -132,13 +135,16 @@ function build_tensorrt_models() {
                 " \
         || (echo "failed streamdiffusion tensorrt"; return 1)
 
-    # TODO: Remove the script download with curl. It should already come in the base image once eliteprox/comfystream#1 is merged.
+    # Depth-Anything-Tensorrt
     docker run --rm -v ./models:/models --gpus all -l TensorRT-engines $AI_RUNNER_COMFYUI_IMAGE \
     bash -c "cd /ComfyUI/models/tensorrt/depth-anything && \
                 python /ComfyUI/custom_nodes/ComfyUI-Depth-Anything-Tensorrt/export_trt.py && \
                 adduser $(id -u -n) && \
                 chown -R $(id -u -n):$(id -g -n) /models" \
     || (echo "failed ComfyUI Depth-Anything-Tensorrt"; return 1)
+
+    # Dreamshaper-8-Dmd-1kstep
+    # TODO: Remove the script download with curl. It should already come in the base image once eliteprox/comfystream#1 is merged.
     docker run --rm -v ./models:/models --gpus all -l TensorRT-engines $AI_RUNNER_COMFYUI_IMAGE \
         bash -c "cd /comfystream/src/comfystream/scripts && \
                  curl -O https://raw.githubusercontent.com/pschroedl/comfystream/refs/heads/10_29/build_trt/src/comfystream/scripts/build_trt.py && \
